filesharingpro/requirements.txt
Django==4.2.17
django-cors-headers==3.7.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2024.2
cryptography